Rewrited socket handling to avoid broken pipe errors.
git-svn-id: https://pl3.projectlocker.com/Tronic/Seraphp/svn/trunk@448 5e58f187-e6ed-c349-b56f-493b30cdff1d

+++ Comm/JsonRpc/JsonRpcProxy.class.php
@@ -452,7 +452,7 @@ class JsonRpcProxy
                 $this->_disconnect($mode);
             }
         }
-        if ($this->_role == 'server') {
+        if ($this->_role == 'server' && $this->_client->getRole() == 'parent') {
             foreach ($this->_fifo as $fifo) {
                 unlink($fifo);
             }
